// src/utils/server/backend/model/fertiscan-db-metadata-inspection-value.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface FertiscanDbMetadataInspectionValue
 */
export interface FertiscanDbMetadataInspectionValue {
  /**
   *
   * @type {number}
   * @memberof FertiscanDbMetadataInspectionValue
   */
  value?: number | null;
  /**
   *
   * @type {string}
   * @memberof FertiscanDbMetadataInspectionValue
   */
  unit?: string | null;
  /**
   *
   * @type {string}
   * @memberof FertiscanDbMetadataInspectionValue
   */
  name?: string | null;
  /**
   *
   * @type {boolean}
   * @memberof FertiscanDbMetadataInspectionValue
   */
  edited?: boolean | null;
}